+++ js/random-int.js
@@ -1,0 +1,9 @@
+//https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+function getRandomInt(min, max) {
+  if (min >=0 && min < max){
+    return Math.floor(Math.random() * (Math.floor(max) - Math.floor(min) + 1)) + Math.floor(min);
+  }
+  return 'Error @#$%^ !!!';
+}
+
+export {getRandomInt};
